Restructure build logic
