+++ rewrite-groovy/build.gradle.kts
@@ -1,7 +1,9 @@
-//plugins {
+plugins {
 //    // temporary, just to write source files to test the limits of Groovy syntax
 //    groovy
-//}
+    id("org.openrewrite.java-library")
+    id("org.openrewrite.maven-publish")
+}
 
 dependencies {
     api(project(":rewrite-java"))
