+++ rewrite-java-11/build.gradle.kts
@@ -1,5 +1,6 @@
 plugins {
-    id("nebula.integtest") version "7.0.9" apply false
+    id("org.openrewrite.java-library")
+    id("org.openrewrite.maven-publish")
 }
 
 apply(plugin = "nebula.integtest-standalone")
@@ -25,7 +26,7 @@ tasks.withType<JavaCompile> {
     sourceCompatibility = JavaVersion.VERSION_11.toString()
     targetCompatibility = JavaVersion.VERSION_11.toString()
 
-    options.release.set(null as? Int) // remove `--release 8` set in root gradle build
+    options.release.set(null as? Int) // remove `--release 8` set in `org.openrewrite.java-base`
     options.compilerArgs.addAll(
         listOf(
             "--add-exports", "jdk.compiler/com.sun.tools.javac.code=ALL-UNNAMED",
