+++ rewrite-protobuf/build.gradle.kts
@@ -1,3 +1,8 @@
+plugins {
+    id("org.openrewrite.java-library")
+    id("org.openrewrite.maven-publish")
+}
+
 // run manually with -x compileKotlin when you need to regenerate
 tasks.register<JavaExec>("generateAntlrSources") {
     main = "org.antlr.v4.Tool"
