+++ rewrite-test/build.gradle.kts
@@ -1,3 +1,8 @@
+plugins {
+    id("org.openrewrite.java-library")
+    id("org.openrewrite.maven-publish")
+}
+
 dependencies {
     api(project(":rewrite-core"))
     api(project(":rewrite-gradle"))
